+++ R/mrca_prior_to_xml_tracelog.R
@@ -14,10 +14,20 @@
 #' @author Richèl J.C. Bilderbeek
 #' @export
 mrca_prior_to_xml_tracelog <- function(
-  clock_models,
-  mrca_prior,
-  tipdates_filename = NA
+  inference_model,
+  clock_models = "deprecated",
+  mrca_prior = "deprecated",
+  tipdates_filename = "deprecated"
 ) {
+  testthat::expect_equal(clock_models, "deprecated")
+  testthat::expect_equal(mrca_prior, "deprecated")
+  testthat::expect_equal(tipdates_filename, "deprecated")
+
+  # Do not be smart yet
+  clock_models <- list(inference_model$clock_model)
+  mrca_prior <- inference_model$mrca_prior
+  tipdates_filename <- inference_model$tipdates_filename
+
   testit::assert(beautier::is_mrca_prior(mrca_prior))
   if (length(mrca_prior) == 1 &&
       beautier::is_one_na(mrca_prior) &&