+++ R/mrca_priors_to_xml_tracelog.R
@@ -18,17 +18,8 @@ mrca_priors_to_xml_tracelog <- function(
   mrca_priors,
   tipdates_filename = NA
 ) {
-  testit::assert(beautier::are_mrca_priors(mrca_priors))
-  text <- NULL
-
-  for (mrca_prior in mrca_priors) {
-    text <- c(text,
-      beautier::mrca_prior_to_xml_tracelog(
-        clock_models = clock_models,
-        mrca_prior = mrca_prior,
-        tipdates_filename = tipdates_filename
-      )
-    )
-  }
-  text
+  stop(
+    "'mrca_priors_to_xml_tracelog' is deprecated",
+    "use 'mrca_prior_to_xml_tracelog' instead"
+  )
 }